chore(rename dir): conform with rename of user/models
