+++ src/modules/user/repositories/IUserRepository.ts
@@ -1,4 +1,4 @@
-import { User } from '../model/User';
+import { User } from '../models/User';
 
 export interface IUserRepository {
   findById(id: string): Promise<User | undefined>;
